+++ src/main/java/world/bentobox/bentobox/util/yaml/YmlCommentParser.java
@@ -1,4 +1,4 @@
-package world.bentobox.bentobox.util;
+package world.bentobox.bentobox.util.yaml;
 
 import java.io.BufferedReader;
 import java.io.IOException;
